requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
asyncpg==0.30.0
alembic==1.16.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-multipart==0.0.20
httpx==0.28.1
apscheduler==3.11.0
pgvector==0.4.1
pymilvus==2.6.0
pypdf==6.0.0
fpdf2==2.8.3
beautifulsoup4==4.13.3
playwright==1.51.0
cryptography==46.0.1
aiofiles==24.1.0
python-telegram-bot==21.6
